Add challenges

+++ src/main/kotlin/challenges/Day01.kt
@@ -12,12 +12,12 @@ class Day01 : Challenge<Int> {
     override fun inputMap(input: List<String>): List<Int> = input
         .map(String::toInt)
 
-    override fun part1(input: List<Int>): Int = input
+    override fun part1(input: List<Int>): Long = input
         .filterWithPrevious(::filterIfIncremented)
-        .count()
+        .count().toLong()
 
-    override fun part2(input: List<Int>): Int = input
+    override fun part2(input: List<Int>): Long = input
         .windowed(3, 1, transform = List<Int>::sum)
         .filterWithPrevious(::filterIfIncremented)
-        .count()
+        .count().toLong()
 }
